bug 3232180
Changed the code that replaces markers to be able to deal with null replacements, meaning that nothiing will be replacing the marker and run the query as is. Now a null just removes the marker, or more accurately, replaces it with an empty string.

Checking in the fix and the jUnit that proves the fix.